Add a way to get he Windows 10 accent from windows APIs

+++ Shared/AccentColor.cs
@@ -8,10 +8,10 @@ namespace BlazorElectronToolbar.Shared
 {
     public class AccentColor
     {
-        public int A { get; set; }
-        public int R { get; set; }
-        public int G { get; set; }
-        public int B { get; set; }
+        public byte A { get; set; }
+        public byte R { get; set; }
+        public byte G { get; set; }
+        public byte B { get; set; }
 
         public override string ToString()
         {
